+++ src/cmdClientConfigure.go
@@ -4,6 +4,7 @@ import "os"
 
 type clientConfigureCmd struct {
 	AMS     clientConfigureAMSCmd     `command:"ams" subcommands-optional:"true" description:"change which clusters prometheus points at"`
+	Tools   clientConfigureToolsCmd   `command:"tools" subcommands-optional:"true" description:"add graph monitoring for AMS for asbenchmark"`
 	Jupyter clientConfigureJupyterCmd `command:"jupyter" subcommands-optional:"true" description:"add language kernels to jupyter"`
 	VSCode  clientConfigureVSCodeCmd  `command:"vscode" subcommands-optional:"true" description:"add languages to VSCode"`
 	Trino   clientConfigureTrinoCmd   `command:"trino" subcommands-optional:"true" description:"change aerospike seed IPs for trino"`